requirements.txt
pandas 
sqlalchemy
python-dotenv
ipykernel
cx_Oracle